requirements.txt
# Core ML Libraries
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# ECG Data Processing
wfdb>=3.4.0
scipy>=1.7.0

# Image Transformation for 2D models
pyts>=0.12.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.62.0

# Utilities
Pillow>=8.3.0
h5py>=3.5.0

# Optional: For better performance
# tensorboard>=2.7.0
# opencv-python>=4.5.0